+++ CHANGELOG.md
@@ -8,6 +8,9 @@ Inspired from [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ### Features
 
+- Add support for agent search in pariwise comparison ([#693](https://github.com/opensearch-project/dashboards-search-relevance/pull/693))
+- Add Search Pipeline to Search Configuration Detail page ([#699](https://github.com/opensearch-project/dashboards-search-relevance/issues/699))
+
 ### Enhancements
 
 * Filter on Experiment creation page the list of Judgments to COMPLETED status only. ([#674](https://github.com/opensearch-project/dashboards-search-relevance/pull/674))